services/python-app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
kafka-python==2.0.2
pydantic==1.10.14
python-multipart==0.0.9
